section remain

+++ backend/audiobook/books/models.py
@@ -87,7 +87,7 @@ class Chapter(models.Model):
 
 class Section(models.Model):
     chapter = models.ForeignKey(Chapter, on_delete=models.CASCADE, related_name='sections')
-    section_number = models.PositiveIntegerField()
+    section_number = models.DecimalField(decimal_places=1,max_digits=10)
     section_name = models.CharField(max_length=200)
     section_image = models.ImageField(upload_to='section_images/', blank=True, null=True)
     created_at = models.DateTimeField(auto_now_add=True)
